requirements.txt
streamlit
langchain
langchain-core
langchain-community
langchain-google-genai>=2.0.0
google-generativeai>=0.7.0
python-dotenv